Draft work
Was trying to find the direction we are facing and get the camera angle, so when passing diagonally across different thresholds e.g. x and z, we can create an obj for the z and x threshold and make one diagonally between them too.

But realised we don't actually need to know the camera angle. We have specific vector points which dictate whether we passed by a threshold or not, so we only need to know what position we are currently in. To know whether to make one diagonally, we will do that manually and decide that based on whether 2 threshold points are perpendicular (at right angles).