Added state, refactoring and find threshold based on current position
Currently implementing logic where we find the threshold point based on the current cube we are occupying - idea is to iterate through all the created cubes, select the cube we are in (where the cords match) and find that's cube threshold point.

If there are any cubes ahead/behind etc we also try to find those cubes and use their threshold points to know whether or not we need to create new cube objs or whether ones exist already and we can just ignore that option.

Also added state, so we can store the threshold points.