feat: MVP done
App now creates cubes whatever direction we are moving whether we are moving forward (positive z) actually moving forward or e.g. turning camera 180 degrees and moving forward (positive z) backwards.

If we move in any diagonals, it also creates cubes in that space.

Had an epiphany whilst on jury service on how I can do this. Realised I can do it in a much easier way. There's 2 things I need:

1. A way to create cubes whenever we hit a threshold whether that is forward, to the side or diagonal
2. A way to detect whether or not a cube has been created at the point when we crossed the threshold

The key is keeping track of the current cube we are in. By doing this, whenever we hit a threshold, we can check if a cube has been created in that space and decide whether to make one there or not.

So instead of having to work out what direction I am travelling in, whether I am moving up or down, all I need to know is what cube I am occupying and with that, whenever I hit a threshold, find out if a cube already exists there or not.

With these 2 simples solutions, I was able to complete my work in like 2hrs!!!! Amazing!!!!

Biggest takeaway: revise and relook at our solution and what we are trying to accomplish. I was making a super complicated solution, when really, all I needed was 2 simple things. Once I had those things in place, everything was pretty much done.